statistics!

+++ bot/plugins/surf.py
@@ -3,6 +3,7 @@ import outlet
 from outlet import errors, String
 from furl import furl
 import discord
+import os
 
 from random import randrange
 
@@ -146,13 +147,14 @@ class Plugin(outlet.Plugin):
             return "Command failed. Try again, and if it still doesn't work, tell reshanie#7510"
 
     async def on_message_delete(self, message):
-        self.log.debug("message delete")
+        if not os.environ.get("RSURF_DEV", False):
+            self.log.debug("message delete")
 
-        ps = self.db.query(self.PrivateServer).filter_by(message_id=message.id).first()  # check if message is PS
-        if ps:  # private server message was deleted
-            self.log.info("Private server message was deleted from list")
+            ps = self.db.query(self.PrivateServer).filter_by(message_id=message.id).first()  # check if message is PS
+            if ps:  # private server message was deleted
+                self.log.info("Private server message was deleted from list")
 
-            self.db.delete(ps)  # remove PS from database on delete
+                self.db.delete(ps)  # remove PS from database on delete
 
-            self.log.debug("comitting db")
-            self.db.commit()
+                self.log.debug("comitting db")
+                self.db.commit()
